refactor: CarTest 분리
